feat(common): adding example plugin
