+++ src/@aui/app/Dashboard.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { PluginLoader } from "./plugins/PluginLoader";
+
+export const Dashboard = () => {
+    return (
+        <PluginLoader />
+    )
+}
