--- src/@aui/app/plugins/myTestPlugin/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const MyTestPlugin = () => {
-    return (
-        <div>My Test Plugin</div>
-    )
-}
-
-export default () => <MyTestPlugin />
