+++ src/@aui/app/plugins/nodeGroups/index.ts
@@ -1,0 +1,1 @@
+export {default} from './NodeGroups'
