+++ src/@aui/app/plugins/processGroups/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ProcessGroups'
